feat: modules tasks completed
