+++ src/modules/esm.mjs
@@ -1,0 +1,46 @@
+import path from 'path';
+import { release, version } from 'os';
+import { createServer as createServerHttp } from 'http';
+import { fileURLToPath } from 'url';
+import fs from 'fs/promises';
+import './files/c.js';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+
+const random = Math.random();
+
+let unknownObject;
+
+const readJsonFile = async (filePath) => {
+    const content = await fs.readFile(filePath, 'utf8');
+    return JSON.parse(content);
+};
+
+if (random > 0.5) {
+    unknownObject = await readJsonFile(path.join(__dirname, 'files', 'a.json'));
+} else {
+    unknownObject = await readJsonFile(path.join(__dirname, 'files', 'b.json'));
+}
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+console.log(`Path to current file is ${__filename}`);
+console.log(`Path to current directory is ${__dirname}`);
+
+const myServer = createServerHttp((_, res) => {
+    res.end('Request accepted');
+});
+
+const PORT = 3000;
+
+console.log(unknownObject);
+
+myServer.listen(PORT, () => {
+    console.log(`Server is listening on port ${PORT}`);
+    console.log('To terminate it, use Ctrl+C combination');
+});
+
+export { unknownObject, myServer };
